put commands in commands
